All Position types seem to work correctly now. Work around spatial search bug.
findContaining() doesn't work in a Module when there are overlapping
functions. Fixed a bug in symbol.cpp to deduplicate symbols, and now
the only cases are ORIG/ORIG_nocancel overlaps. Hack around this for
now with a local search within the 5 nearest functions. Should use a
spatial data structure that supports overlapping intervals.