Fix X86_64 port.
We need to make a marker link for X86_64 too for a newer version of
glibc that has a section called __libc_IO_vtables. For this version, it
has a link called __stop___libc_IO_vtables which points to the start
address of that section + its size (and we only make data links that
point to inside data sections now).