Handle weak reference for AARCH64.
The LogCalls pass is also fixed to resolve to the address of the Conductor
used in loader. However, this still does not work on some platforms
where a memory object allocated in the loader could end up being
destroyed in the target program.

For X86_64, this is not supported yet. We need to first create
DataVariables for all relocations in data regions.