Improve jump table detection, spotting ptmalloc_init.part.14's table.
We now consider that if any sequence of blocks can lead to a register
having an lea instruction, then presumably this is a jump table and the
other options (all bare registers in this case) are from an incomplete
call graph, since jump tables are not yet known.

Also optimizing some multiple parents cases.