Properly redirect calls to PLTs to the newly-generated PLT.
Currently there's a rather hacky hard-coding of the address of the PLT, which
should be removed at some point. However, since the .plt is a text section,
it doesn't get assigned an address until much later on in the pipeline and I
couldn't find a nice place to insert it.

(written by ethereal, blame me, not dwk, for bugs in this code)